iSCSI Console v1.2.2

+++ ISCSI/PDU/RejectPDU.cs
@@ -1,4 +1,4 @@
-/* Copyright (C) 2012-2015 Tal Aloni <tal.aloni.il@gmail.com>. All rights reserved.
+/* Copyright (C) 2012-2016 Tal Aloni <tal.aloni.il@gmail.com>. All rights reserved.
  * 
  * You can redistribute this program and/or modify it under the terms of
  * the GNU Lesser Public License as published by the Free Software Foundation,
@@ -21,7 +21,7 @@ namespace ISCSI
 
         public RejectPDU() : base()
         {
-            OpCode = (byte)ISCSIOpCodeName.Reject;
+            OpCode = ISCSIOpCodeName.Reject;
             Final = true;
             InitiatorTaskTag = 0xFFFFFFFF;
         }
