+++ ISCSI/PDU/SCSIResponsePDU.cs
@@ -1,4 +1,4 @@
-/* Copyright (C) 2012-2015 Tal Aloni <tal.aloni.il@gmail.com>. All rights reserved.
+/* Copyright (C) 2012-2016 Tal Aloni <tal.aloni.il@gmail.com>. All rights reserved.
  * 
  * You can redistribute this program and/or modify it under the terms of
  * the GNU Lesser Public License as published by the Free Software Foundation,
@@ -29,7 +29,7 @@ namespace ISCSI
 
         public SCSIResponsePDU() : base()
         {
-            OpCode = (byte)ISCSIOpCodeName.SCSIResponse;
+            OpCode = ISCSIOpCodeName.SCSIResponse;
             Final = true;
         }
